Moved connection and session to their own file
